django_movie/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.4
defusedxml==0.7.1
Django==4.2
django-allauth==0.61.1
django-ckeditor==6.7.1
django-environ==0.11.2
django-filter==23.5
django-js-asset==2.2.0
django-modeltranslation==0.18.11
django-recaptcha3==0.4.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
docker==7.0.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
gunicorn==21.2.0
